+++ Assets/Scripts/Effects/TakeDamageEffect.cs
@@ -9,7 +9,7 @@ public class TakeDamageEffect : InstantCharacterEffect
     [Header("Damage")]
     public float physicalDamage = 0; //usually split into standard, strike, slash and piece
     public float magicDamage = 0;
-    private float finalDamageDealt = 0;
+    private int finalDamageDealt = 0;
 
     [Header("Poise")]
     public float poiseDamage = 0;
